Ft 164489790 Delete comment
