+++ routes/api/articles.js
@@ -1,5 +1,6 @@
 import express from 'express';
 import articleController from '../../controllers/article';
+import articleCommentController from '../../controllers/articlecomment';
 import secureRoute from '../../middlewares/tokenValidation';
 
 const router = express.Router();
@@ -9,8 +10,9 @@ router.get('/');
 router.delete('/:slug');
 router.put('/:slug');
 
-router.post('/:slug/comments', secureRoute, articleController.addComment);
-router.get('/:slug/comments', secureRoute, articleController.getComments);
-router.put('/:slug/comments/:commentid', secureRoute, articleController.updateComment);
+router.post('/:slug/comments', secureRoute, articleCommentController.addComment);
+router.get('/:slug/comments', secureRoute, articleCommentController.getComments);
+router.put('/:slug/comments/:commentid', secureRoute, articleCommentController.updateComment);
+router.delete('/:slug/comments/:commentid', secureRoute, articleCommentController.deleteComment);
 
 export default router;
